add Comedi DAC control
git-svn-id: https://svn.code.sf.net/p/opticsbenchui/code@32 19f4beda-a93a-4e52-9a4f-25590d2c5071

+++ AcquisitionSequence.h
@@ -80,7 +80,7 @@ class AcquisitionSequence
 
   // Counter Comedi attributes
   int    comediValue;
-  float  comediData;
+  double comediData;
   int    comediOutput;
 
   // SLM attributes
