+++ Comedi.h
@@ -31,7 +31,7 @@ class Comedi : public QObject
   virtual bool connectComedi(QString newcomedi) = 0;
   virtual bool resetComedi(QString newcomedi) = 0;
   virtual bool setComediValue(QString newcomedi, int output, void *value) = 0;
-  virtual bool getComediValue(QString newcomedi, int output, float &value) = 0;
+  virtual bool getComediValue(QString newcomedi, int output, double &value) = 0;
   virtual bool updateDBValues(QString newcomedi) = 0;
   virtual void setDbPath(QString _path) = 0;
